requirements.txt
streamlit==1.35.0
numpy==1.26.4
pillow==10.3.0

torch==2.2.2
torchvision==0.17.2

scikit-learn==1.4.2